chore(Drip): fix schema
